[Bronze V] Title: 이상한 기호, Time: 180 ms, Memory: 17704 KB, Score: 100 point - DashHub

+++ BOJ/Bronze/15964. 이상한 기호/Main.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		long a = sc.nextLong();
+		long b = sc.nextLong();
 		long res = (a + b) * (a - b);
 		System.out.println(res);
 	}
